Renamed the "smesh.*" framework files to adhere to the naming convention of this library code; certain edits were made to the files for this to work.
Added a global pointer to the UID factory of the "smesh" code.

Modified the Makefile to do proper linking of all the code. (This is not
expected to use any of the "smesh" code at this point, and it will not work
in plain C without wrapper functions, which I should add at some point.)
IN 2025/10/22 "master"